Update resolution example to use the new Resolver (#838)
Updates the resolution example to use the new Resolver.
