+++ bindings/wasm/examples/README.md
@@ -44,7 +44,7 @@ The following examples are currently available:
 |  1  | [create_did](src/create_did.js)           | Generates and publishes a DID Document, the fundamental building block for decentralized identity.                                                                                                                                         |
 |  2  | [manipulate_did](src/manipulate_did.js)   | Add verification methods and service endpoints to a DID Document and update an already existing DID Document.                                                                                                                              | |
 |  3  | [resolve_history](src/resolve_history.js) | Advanced example that performs multiple updates and demonstrates how to resolve the DID Document history to view them.                                                                            |
-|  4  | [resolution](src/resolution.js)           | Resolves an existing DID to return the latest DID Document.                                                                                                                                                                                |
+|  4  | [resolution](src/resolve_did.js)          | Resolves an existing DID to return the latest DID Document.                                                                                                                                                                                |
 |  5   | [key_exchange](src/key_exchange.js) | Demonstrates Elliptic-curve Diffie-Hellman (ECDH) cryptographic key exchange with DID Documents. | |
 |  6   | [private_tangle](src/private_tangle.js)   | Showcases the same procedure as `create_did`, but on a private tangle - a locally running hornet node.                                                                                                                                     |
