+++ bindings/wasm/examples/src/tests/node/resolution.js
@@ -1,9 +1,9 @@
-import {resolution} from "../../resolution";
+import {resolveDID} from "../../resolve_did";
 import {CLIENT_CONFIG} from "../../config";
 
 // Only verifies that no uncaught exceptions are thrown, including syntax errors etc.
 describe("Test node examples", function () {
     it("Resolution", async () => {
-        await resolution(CLIENT_CONFIG);
+        await resolveDID(CLIENT_CONFIG);
     });
 })
